fix: moved dropdown component from Fields to section fields for contact page
