+++ frontend/src/components/SectionFields.vue
@@ -19,12 +19,63 @@
         class="grid min-h-[28px] flex-1 items-center overflow-hidden text-base"
       >
         <div
-          v-if="field.read_only && field.type !== 'checkbox'"
+          v-if="
+            field.read_only && !['checkbox', 'dropdown'].includes(field.type)
+          "
           class="flex h-7 cursor-pointer items-center px-2 py-1 text-gray-600"
         >
           <Tooltip :text="__(field.tooltip)">
             <div>{{ data[field.name] }}</div>
           </Tooltip>
+        </div>
+        <div v-else-if="field.type === 'dropdown'">
+          <NestedPopover>
+            <template #target="{ open }">
+              <Button
+                :label="data[field.name]"
+                class="dropdown-button flex w-full items-center justify-between rounded border border-gray-100 bg-gray-100 px-2 py-1.5 text-base text-gray-800 placeholder-gray-500 transition-colors hover:border-gray-200 hover:bg-gray-200 focus:border-gray-500 focus:bg-white focus:shadow-sm focus:outline-none focus:ring-0 focus-visible:ring-2 focus-visible:ring-gray-400"
+              >
+                <div class="truncate">{{ data[field.name] }}</div>
+                <template #suffix>
+                  <FeatherIcon
+                    :name="open ? 'chevron-up' : 'chevron-down'"
+                    class="h-4 text-gray-600"
+                  />
+                </template>
+              </Button>
+            </template>
+            <template #body>
+              <div
+                class="my-2 space-y-1.5 divide-y rounded-lg border border-gray-100 bg-white p-1.5 shadow-xl"
+              >
+                <div>
+                  <DropdownItem
+                    v-if="field.options?.length"
+                    v-for="option in field.options"
+                    :key="option.name"
+                    :option="option"
+                  />
+                  <div v-else>
+                    <div class="p-1.5 px-7 text-base text-gray-500">
+                      {{ __('No {0} Available', [field.label]) }}
+                    </div>
+                  </div>
+                </div>
+                <div class="pt-1.5">
+                  <Button
+                    variant="ghost"
+                    class="w-full !justify-start"
+                    :label="__('Create New')"
+                    @click="field.create()"
+                  >
+                    <template #prefix>
+                      <FeatherIcon name="plus" class="h-4" />
+                    </template>
+                  </Button>
+                </div>
+              </div>
+            </template>
+          </NestedPopover>
         </div>
         <FormControl
           v-else-if="field.type == 'checkbox'"
@@ -109,13 +160,21 @@
         class="h-4 w-4 shrink-0 cursor-pointer text-gray-600 hover:text-gray-800"
         @click="field.link(data[field.name])"
       />
+      <EditIcon
+        v-if="field.type === 'link' && field.edit && data[field.name]"
+        class="size-3.5 shrink-0 cursor-pointer text-gray-600 hover:text-gray-800"
+        @click="field.edit(data[field.name])"
+      />
     </div>
   </FadedScrollableDiv>
 </template>
 
 <script setup>
+import NestedPopover from '@/components/NestedPopover.vue'
+import DropdownItem from '@/components/DropdownItem.vue'
 import FadedScrollableDiv from '@/components/FadedScrollableDiv.vue'
 import ArrowUpRightIcon from '@/components/Icons/ArrowUpRightIcon.vue'
+import EditIcon from '@/components/Icons/EditIcon.vue'
 import Link from '@/components/Controls/Link.vue'
 import UserAvatar from '@/components/UserAvatar.vue'
 import { usersStore } from '@/stores/users'
@@ -188,7 +247,8 @@ function evaluate(code, context = {}) {
 :deep(.form-control input:not([type='checkbox'])),
 :deep(.form-control select),
 :deep(.form-control textarea),
-:deep(.form-control button) {
+:deep(.form-control button),
+.dropdown-button {
   border-color: transparent;
   background: white;
 }
